requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
pydantic[email]==2.10.2
aiofiles==24.1.0
Pillow==11.0.0
reportlab==4.2.5
